Fixed push navigation issue
