+++ android/app/src/main/java/app/esteem/mobile/MainApplication.java
@@ -12,7 +12,6 @@ import com.reactnative.ivpusic.imagepicker.PickerPackage;
 import com.microsoft.appcenter.reactnative.push.AppCenterReactNativePushPackage;
 import com.entria.views.RNViewOverflowPackage;
 import com.oblador.vectoricons.VectorIconsPackage;
-import com.avishayil.rnrestart.ReactNativeRestartPackage;
 import com.BV.LinearGradient.LinearGradientPackage;
 import com.lugg.ReactNativeConfig.ReactNativeConfigPackage;
 import com.microsoft.appcenter.reactnative.crashes.AppCenterReactNativeCrashesPackage;
@@ -22,7 +21,6 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
-import com.facebook.react.BuildConfig;
 
 import java.util.Arrays;
 import java.util.List;
@@ -54,7 +52,6 @@ public class MainApplication extends Application implements ReactApplication {
             new AppCenterReactNativePushPackage(MainApplication.this),
             new RNViewOverflowPackage(),
             new VectorIconsPackage(),
-            new ReactNativeRestartPackage(),
             new LinearGradientPackage(),
             new ReactNativeConfigPackage(),
             new AppCenterReactNativeCrashesPackage(MainApplication.this, getResources().getString(R.string.appCenterCrashes_whenToSendCrashes)),
